SDK-1659: Move token validation to profile service, move profile tests from yoti_client_test into (profile/)service_test
