+++ profile/service.go
@@ -24,6 +24,10 @@ func getProfileEndpoint(token, sdkID string) string {
 // GetActivityDetails requests information about a Yoti user using the one time
 // use token generated by the Yoti login process. Don't call this directly, use yoti.GetActivityDetails
 func GetActivityDetails(httpClient requests.HttpClient, token, clientSdkId, apiUrl string, key *rsa.PrivateKey) (activity ActivityDetails, err error) {
+	if len(token) < 1 {
+		return activity, errors.New("Invalid Token")
+	}
+
 	var decryptedToken string
 	decryptedToken, err = cryptoutil.DecryptToken(token, key)
 	if err != nil {
